32.1. Product CRUD --> yarn add bootstrap@5 --dev
I've mentioned a few times that Stimulus has a sister technology called Turbo... both of which live under this brand called "Hotwire", which, as we've learned, is all about returning HTML from your server.

Turbo: You'll Write Less Custom JavaScript
We're going to discuss Turbo in the next tutorial in this series. But in a nutshell, Turbo allows you to turn all of the clicks and form submits on your site into Ajax calls instantly & automatically. And it has several other neat super powers.

Right now, on the cart page, when we remove an item, we did some extra work to submit the delete form via Ajax and reload the cart area also via Ajax. Once you start using Turbo, you absolutely can still do stuff like this... but you'll find that it's less necessary. In this case, if Turbo were active, after confirming in the modal, we could just let the form submit normally... which would automatically happen via Ajax. My point is: Turbo will allow us to have a slick user interface... while writing less custom JavaScript. But if you ever want to do something extra, you are completely free to write custom JavaScript.

Let's Build an AJAX Form Modal!
In that spirit, after talking with a few of you wonderful people, I thought it might be good to show how we could submit a full form via Ajax, including handling validation errors and reloading part of the page after that Ajax call finishes.
So here's the plan: we're going to generate a new product admin section. Then, on the list page. we'll add a "new product" button that, on click opens a modal with a form inside. We'll submit that form via Ajax in the modal, show validation errors in the modal and, finally, on success, reload the product list on the page via another Ajax call. It's going to be an epic example of Stimulus. We're also going to do part of this using jQuery... just in case you prefer using it over vanilla JavaScript.

Upgrading to Bootstrap 5
So let's get going! To start, find your terminal.
The project is currently using Bootstrap version 4. I'm going to upgrade to Bootstrap 5 because I like it's JavaScript components better. Run:

yarn add bootstrap@5 --dev
At this exact moment, Bootstrap 5 is only in beta. So I'll select that version.

Peer Dependency Warnings?
See this "peer dependency" warning? If you ever see these errors - like up here - and they mention Webpack or Babel... it's probably fine. This happens because Encore handles so much stuff for us. These libraries that it's complaining about are installed... but they're installed by Encore directly.

But in this case, this popper thing is going to be a problem... but we can wait to see what error it causes. Bootstrap 5 does change some class names versus Bootstrap 4... but it's minor enough that I'm going to ignore it. If you look at our site, the page still looks just fine. You will notice a few things that don't look right on our form... but we won't worry about fixing those.
Oh, and by the way, if you haven't already after the last chapter, make sure you stop the webpack-bundle-analyzer and restart Encore with:

yarn watch