32.5. Product CRUD
 Let's improve that a tiny bit. Open up this page's template, which is templates/product_admin/index.html.twig. On top, add a <div> class container-fluid and mt-4 for some margin. All the way at the bottom, add the ending div.
Copy this... because we need it in all of our templates. Open edit, do that same thing... add the closing div... new.html.twig... and finally show.html.twig.
When you're done, refresh the list page again. Okay: it's slightly better. It could still use some margin over here, but it's good enough for now.
Click to edit a product. Ah!

  Error: class Category could not be converted to string.

Rude! This is because the form is trying to make a category select drop down... and it needs to know what text to use for each Category option.