32.6. Product CRUD
Adding __toString to Make the Select Field Work
Go into src/Entity/Category.php and, anywhere in here - I'll put it at the bottom - add a public function __toString() method that returns $this->name. I'm going to cast that to a string... just in case the name is null.
Oh, and while we're thinking about the form, I want to make it a bit smaller. In src/Form/ProductType.php, the form contains every field. To make life simpler, remove brand, weight, stockQuantity, imageFilename, and also colors.
Very nice.
Refresh the edit page now and... it works! If we change something and hit update - yes our buttons do need some styling - that works too.
Oh, but there's one more change I want to make: this will help our example. Back over in ProductAdminController::index(), change the query to sort the newest on top. Do that by changing this findAll() to findBy(), pass it an empty criteria - so it still returns everything - and then sort by id DESC. You could also use a createdAt column if you want.
Head over and refresh now. Perfect: the highest ids are on top.
Ok: our setup is complete! Next: let's create an "add" button right here on the list page that, on click, opens a Bootstrap modal with our form inside. We'll accomplish that with a Stimulus controller and an Ajax call.