33.3. Opening a Modal
Attaching the Button Action
Step two: on click, we want to open a modal. This means we need to add an action to the button.
Over in our template, add data-action= the name of our controller - modal-form - pound sign, and let's call the method openModal.
Copy that, head into the controller, rename connect() to openModal() and add the event argument in case we need it. Inside, console.log(event).
If we refresh now... and click. We're on a roll!