33.7. Opening a Modal
Opening the Modal
Head back to our controller and remove the console.log(). Now say: const modal = new Modal().
What we need to pass here is the DOM element that holds the modal template. In other words, this element right here. How can we find that from inside of our controller? By using a target of course!
Back in _modal.html.twig, all the way up on the top level element, add a target: data- - the name of our controller - modal-form - -target= and call the new target modal.
This does make the modal template a bit specific to this one Stimulus controller. But I'm okay with that. If we need to make this same element a target for a different controller in the future... we can totally do that! We can add as many target attributes as we need.
Copy the target name and head back to the controller. Declare it with static targets = [] an array with modal inside. Careful with the typing on targets: I'll regret that mistake.
Anyways, now we can say new Modal(this.modalTarget).
That creates a new Modal object... but doesn't actually open it yet. To do that, say modal.show().
Time to take it for a test drive! Move over, refresh and click. Ah! An error!

  Cannot read property classList of undefined.

It's coming from Bootstrap. It's not perfectly clear what's happening... but the "undefined" is very telling. It makes me wonder if my target isn't being seen correctly.