34.1. Loading a Form into the Modal
We're going to load the new product form into the body of this modal. But the header and buttons will still come from _modal.html.twig.

Making _modal.html.twig Customizable
Let's customize those to make more sense. Up on the header, we can say: "add a new product".
Wait, don't do that. I want to try to make this template as reusable as possible for other modals. Instead, let's say {{ modalTitle }}
In index.html.twig, add a second argument to include() and pass modalTitle set to "Add a new product"
Very nice! For the body, use {{ modalContent }}. That's a new variable I'm inventing. But pipe this into the default filter and say "loading..."
In this case, we are not going to pass any modal content, but you could in other situations. We'll replace the loading... in a minute after we make the Ajax call.
For the buttons, hard-code those to some new text: "Cancel" and "Save". We can always make them dynamic later.
Let's make sure we didn't break anything. When I click the button, very nice!