34.2. Loading a Form into the Modal
Passing the Form AJAX URL to the Stimulus Controller
To get the new product form HTML, when the modal opens, we're going to make an Ajax call to an endpoint that will return that HTML.
Head over to src/Controller/ProductAdminController.php and find the new action. This is the endpoint that we're going to make our Ajax request to: we're going to customize this so that it's able to return the full HTML page or just the form partial under a specific condition. We'll do that in a minute.
Copy the route name. As you know, I don't like to hard-code Ajax URLs in my Stimulus controllers... and I really don't want to do that in this case because I want the controller to be reusable for other forms on our site.
And so, we'll do what we've done several times before: pass the URL into the controller as a value. Add static values = {} and create a value called, how about formUrl, which will be a String.
Then, down in openModal, console.log(this.formUrlValue).
In the template, on stimulus_controller, add a second argument so that we can pass the formUrl value set to path() and the route name: product_admin_new.
Try it: refresh, click and... got it! There's the URL.