34.4. Loading a Form into the Modal
Making the Ajax Call
Now, down in the method, remove the console.log() and make the Ajax call with $.ajax() and pass it this.formUrlValue.
That will make the Ajax call... but will do absolutely nothing with the result. What we need to do is take the HTML from the Ajax call and, if you look at _modal.html.twig, put it inside the modal-body element. That means we need a new target.
Right here, add data-modal-form-target= and let's call this one modalBody.
Copy that, go back to the controller, and set this up as a second target.
In openModal() use that: this.modalBodyTarget.innerHTML equals, await $.ajax()... because jQuery's Ajax function returns a Promise. And, of course, my Webpack build is mad because we need to make openModal() async.
Our Ajax call is still going to return the HTML for the entire page... but let's at least see if it works.
Move over, refresh and... awesome! It looks totally wrong because the endpoint returns the full page, but it is working!
Before we fix that, I want to handle one small detail. In our Stimulus controller, at the very top of openModal, add this.modalBodyTarget.innerHTML equals Loading....
That's a minor thing: if we open the modal twice, this will clear the contents before we start the AJAX call... so that we don't temporarily see an old form.