34.5. Loading a Form into the Modal
We probably want to keep the buttons in the modal footer and hide the one that's coming from the form partial. A really easy way to do this is with CSS. Over in your editor, open assets/styles/app.css. All the way at the bottom, we're going to hide any buttons that are inside of the modal body... which has this modal-body class. Do that with .modal-body button and display: none.
This will hide all the buttons for all of the modals on your site. If that's a problem, add a custom class on your modal HTML so you can be more targeted.
When I refresh now... and click the button... it looks perfect!
Okay: we've got our form into the modal. Now we need to make it submit via Ajax inside the modal. Lets do that next!