35.2. Ajax-Submitting an Entire Form
Making the Ajax Call

To make the Ajax call, we need three things... and we already have the first. We need the form field data - we have that with .serialize() and also the URL to submit to & method to use. We can get those last two directly from the form element.
Say $.ajax() and I'll pass it the options format where even the URL is an option. Set that to $form.. Now you might expect me to read the action attribute off of the form. But instead, say .prop('action').
That's slightly different... and bit smarter: this will return the correct action URL even if there is no action attribute... which means that a form should submit back to the current URL.
If you look back at submit-confirm_controller, that time we used this.element.action... which is a property that exists on all form elements. In jQuery, we're asking it to give us that same property.
Repeat this for method set to $form.prop('method').
Finally, for the data, we can say data set to $form.serialize().
So there you have it: a jQuery version of how you can submit a form via Ajax.
Now, this will submit the form... but we want to also do something with the response.
Look at ProductAdminController::new(). If we submit to this URL and the form fails validation, it will automatically re-render our _form.html.twig template. But this time the form will render with errors.
So... great! This means that after the Ajax call is finished, we want to put the returned HTML back onto this.modalBodyTarget. Copy that and say this.modalBodyTarget.innerHTML equals await $.ajax(). Then make submitForm() async.

Debugging the Ajax Error
Moment of truth! Spin over, refresh the page, click the button, but leave the form blank this time. Hit save and... ah! There's an error!
What happened? Let's find out next... and make sure that our form - no matter how or where it's submitted - will always work.