37.1. Ajax Form Response Status Codes
Open the modal and fill out the name and price fields: the only two required fields. I want to see what happens if we submit the form successfully.
And... whoa! That was weird. On success, our controller currently redirects back to /admin/product. When jQuery's Ajax hits a redirect, it follows it and makes a second Ajax request. We can see this in the Network tab: here's the original request and here's the request from the redirect. And because that redirects to the list page, it grabbed the full HTML for that page and jammed it into the modal.
That's... probably not what we want. Instead, I want to close the modal. But to do that, we need to be able to determine - from JavaScript - whether a form submit was successful or had a validation error.
But, at the very least, if we refresh the page, we do see the new product in the list. So, yay! We know the save part is working.

On Success: 204 Empty Response
Over in the controller, if the form is submitted via Ajax, instead of a redirect, what should we return? How about an empty, successful response. We can do that with if $request->isXmlHttpRequest() - or you might instead be looking for a query parameter if you decided to use that strategy - then return new Response() - the one from HttpFoundation - passing it null for the content and 204 for the status code.
204 is a special status code that means:

  This response was successful! Yay! But... I have no content that I want to return.

So: an empty - but successful - response. That'll fix the redirect issue.