37.2. Ajax Form Response Status Codes
On Error: 422 Error Response
But there's still one problem. When the form has validation errors, we're currently returning a 200 success status code: that's what $this->render() uses by default.
In other words, whether the form submit is successful or not, both situations return a successful status code. That... doesn't make our life very easy in JavaScript where we need to figure out which situation we're in! Sure, we could look exactly for a 200 versus 204 status code... but... there's a better way.
What is it? Return an error status code when the form fails validation. This will not only make our life easier in JavaScript... it's also more correct! And it will still work fine if you're submitting a form normally without JavaScript: the error status code won't confuse old browsers or anything like that.
The third argument to render() is a Response object that the HTML from the template will be put into. When this is not passed, a Response object with a 200 status code is automatically created. Let's pass our own: new Response() - the same one from HttpFoundation - passing null for the for the content... because that's going to be replaced by the HTML from the template anyways.
The really important thing is the status code. But this render() method is called both when the form is originally loaded and when it's submitted with invalid data. So we need to figure out which situation we're in. Use the ternary syntax here to say: if $form->isSubmitted(), then use 422 else 200.
This works because - if the form was submitted and was successful - we would already be inside the first if statement... and we would never get down here. So if the form is submitted, we definitely know it's an invalid submit.
The 422 status code means "unprocessable entity". And it's a pretty standard status code for validation errors. As a bonus, doing this on your forms will play super nicely with Stimulus's sister technology Turbo. Oh, and by the way, in Symfony 5.3, there is a new renderForm() shortcut in your controller, which will automatically set the 422 status code for you on error. That'll make this much cleaner.